changeUserList.js -> addUser.js, deleteUser.js
